+++ src/tkc/event_source_manager.c
@@ -126,8 +126,16 @@ uint32_t event_source_manager_get_wakeup_time(event_source_manager_t* manager) {
       }
     }
   } else {
-    wakeup_time = TK_DEFAULT_WAIT_TIME;
+    wakeup_time = manager->min_sleep_time;
   }
 
-  return tk_min(TK_DEFAULT_WAIT_TIME, wakeup_time);
+  return tk_min(manager->min_sleep_time, wakeup_time);
+}
+
+ret_t event_source_manager_set_min_sleep_time(event_source_manager_t* manager, uint32_t sleep_time) {
+  return_value_if_fail(manager != NULL, RET_BAD_PARAMS);
+
+  manager->min_sleep_time = sleep_time;
+
+  return RET_OK;
 }