+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2023/06/17
+  * 增加函数event\_source\_manager\_set\_min\_sleep\_time。
+
 2023/06/15
   * 修复widget\_get\_style在特殊状态下获取普通风格数据出错的问题(感谢智明提供补丁)
   * 修改label\_auto\_adjust\_size调用widget\_resize来修改宽高(感谢智明提供补丁)